Add files via upload
For the reflection of what was created and what ChatGPT created, they were pretty similar; I did use some help from other classmates as well as outside sources (not ChatGPT but YouTube, hehe). For the second part of this if I had only used sources for AI I do not think I could adequately explain the code I had written, it is so easy just to ask AI to write everything for you and not do it or learn it yourself.